runtime/core/requirements.txt
# Core runtime dependencies (minimal).
fastapi>=0.110.0
uvicorn>=0.27.0
PyYAML>=6.0.1
jsonschema>=4.21.0
referencing>=0.28.4
